Split test_cli.py
