+++ tests/test_cli_compile.py
@@ -11,9 +11,9 @@ import pytest
 from piptools._compat.pip_compat import path_to_url
 from piptools.repositories import PyPIRepository
 from piptools.scripts.compile import cli
-from piptools.scripts.sync import cli as sync_cli
 from pip._vendor.packaging.version import parse as parse_version
 from pip import __version__ as pip_version
+from .utils import invoke
 
 
 PIP_VERSION = parse_version(os.environ.get('PIP', pip_version))
@@ -174,25 +174,10 @@ def test_realistic_complex_sub_dependencies(tmpdir):
         assert out.exit_code == 0
 
 
-def _invoke(command):
-    """Invoke sub-process."""
-    try:
-        output = subprocess.check_output(
-            command,
-            stderr=subprocess.STDOUT,
-        )
-        status = 0
-    except subprocess.CalledProcessError as error:
-        output = error.output
-        status = error.returncode
-
-    return status, output
-
-
 def test_run_as_module_compile(tmpdir):
     """piptools can be run as ``python -m piptools ...``."""
 
-    status, output = _invoke([
+    status, output = invoke([
         sys.executable, '-m', 'piptools', 'compile', '--help',
     ])
 
@@ -201,37 +186,6 @@ def test_run_as_module_compile(tmpdir):
     assert output.startswith('Usage:')
     assert 'Compiles requirements.txt from requirements.in' in output
     assert status == 0
-
-
-def test_run_as_module_sync():
-    """piptools can be run as ``python -m piptools ...``."""
-
-    status, output = _invoke([
-        sys.executable, '-m', 'piptools', 'sync', '--help',
-    ])
-
-    # Should have run pip-compile successfully.
-    output = output.decode('utf-8')
-    assert output.startswith('Usage:')
-    assert 'Synchronize virtual environment with' in output
-    assert status == 0
-
-
-def test_sync_quiet(tmpdir):
-    """sync command can be run with `--quiet` or `-q` flag."""
-
-    runner = CliRunner()
-    with runner.isolated_filesystem():
-        with open('requirements.txt', 'w') as req_in:
-            req_in.write('six==1.10.0')
-
-        with mock.patch('piptools.sync.check_call') as check_call:
-            out = runner.invoke(sync_cli, ['-q'])
-            assert out.output == ''
-            assert out.exit_code == 0
-            # for every call to pip ensure the `-q` flag is set
-            for call in check_call.call_args_list:
-                assert '-q' in call[0][0]
 
 
 def test_editable_package(tmpdir):
